+++ pkg/util/base.go
@@ -1,6 +1,8 @@
 package util
 
 import (
+	"fmt"
+	"crypto/md5"
 	"k8s.io/client-go/kubernetes"
 	"k8s.io/client-go/tools/clientcmd"
 	"os"
@@ -27,4 +29,11 @@ func GetClientSetForTest(t *testing.T) *kubernetes.Clientset {
 		t.Errorf("failed to NewForConfig, %++v", err)
 	}
 	return clientset
+}
+
+func GetMd5V2(str string) string {
+    data := []byte(str)
+    has := md5.Sum(data)
+    md5str := fmt.Sprintf("%x", has)
+    return md5str
 }